+++ lib/entities/custom/custom_races.dart
@@ -1,4 +1,3 @@
-
 // A file for exposing all the base generators for creating custom races
 
 // ---------------------- Physical -----------------