+++ lib/src/subtypes/kingdoms/default_kingdom/default_kingdom_type.dart
@@ -1,0 +1,89 @@
+import '../../../entities/kingdoms/kingdom.dart';
+import '../../../generators/base/future_generator.dart';
+import '../../../generators/base/generator.dart';
+import '../../../generators/base/list_item_generator.dart';
+import '../../../generators/base/number_generator.dart';
+import '../../../generators/fixable.dart';
+import '../../../generators/kingdoms/history/kingdom_history_generator.dart';
+import '../../../generators/kingdoms/known_for/kingdom_known_for_generator.dart';
+import '../../../generators/kingdoms/trouble/kingdom_trouble_generator.dart';
+import '../../../generators/settlements/names/dominant_race_name_generator.dart';
+import '../../guilds/guild_manager.dart';
+import '../../guilds/guild_type.dart';
+import '../../races/race.dart';
+import '../../settlements/city/city.dart';
+import '../../settlements/metropolis/metropolis.dart';
+import '../../settlements/settlement_type.dart';
+import '../../settlements/town/town.dart';
+import '../government_types/government_type.dart';
+import '../government_types/government_type_manager.dart';
+import '../kingdom_type.dart';
+
+/// A class that represents the default kingdom type
+class DefaultKingdomType implements KingdomType, Fixable<Kingdom> {
+  const DefaultKingdomType();
+
+  static const _kingdomType = "default";
+  static const _minPopulation = 80000;
+  static const _maxPopulation = 11000000;
+  static const _populationBarrier = 25000;
+
+  @override
+  IGenerator<SettlementType> getCapitalTypeGenerator() => ListItemGenerator([
+        Metropolis(),
+        City(),
+      ]);
+
+  @override
+  IGenerator<GovernmentType> getGovernmentTypeGenerator() =>
+      ListItemGenerator(GovernmentTypeManager.activeGovernmentTypes);
+
+  @override
+  IGenerator<GuildType> getGuildTypeGenerator() =>
+      ListItemGenerator(GuildManager.activeGuildTypes);
+
+  @override
+  IGenerator<String> getHistoryGenerator(String kingdomName) =>
+      KingdomHistoryGenerator(kingdomName);
+
+  @override
+  IGenerator<SettlementType> getImportantSettlementsTypesGenerator() =>
+      ListItemGenerator([
+        City(),
+        Town(),
+      ]);
+
+  @override
+  IGenerator<String> getKnownForGenerator() => KingdomKnownForGenerator();
+
+  @override
+  IGenerator<String> getNameGenerator(Race race) =>
+      DominantRaceNameGenerator(race);
+
+  @override
+  IGenerator<int> getPopulationGenerator() => FutureGenerator(
+        NumberGenerator(_minPopulation, _maxPopulation),
+        (population) => population ~/ 1000 * 1000,
+      );
+
+  @override
+  IGenerator<String> getTroubleGenerator() => KingdomTroubleGenerator();
+
+  @override
+  String getKingdomType() => _kingdomType;
+
+  @override
+  Kingdom getFixed(Kingdom kingdom) {
+    final populations = [
+      ...kingdom.importantSettlements.map((e) => e.population),
+      kingdom.capital.population,
+    ];
+    final populationSum =
+        populations.reduce((value, element) => value + element);
+
+    if (populationSum + _populationBarrier >= kingdom.population) {
+      return kingdom.copyWith(population: kingdom.population * 10);
+    }
+    return kingdom;
+  }
+}