+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Table of Contents <!-- omit in toc -->
 
+- [0.8.0](#080)
 - [0.7.0](#070)
 - [0.6.0](#060)
 - [0.5.2](#052)
@@ -14,6 +15,18 @@
 - [0.1.1](#011)
 - [0.1.0](#010)
 - [0.0.0](#000)
+
+## 0.8.0
+
+- Added kingdoms
+
+- Switched settlements for kingdoms in world entity
+
+- Added government types
+  - Federation
+  - Monarchy
+  - Republic
+  - Theocracy
 
 ## 0.7.0